==> notebooks/00-summarize-fb-page-info.ipynb <==
# Summarize general info for Facebook pages

This notebook combines the `pages-info.csv` and `all-partisan-sites.csv` files to produce a file with four attributes for each Facebook page:

- `page_id`
- `political_category`
- `page_name`
- `fan_count`

import pandas as pd

page_info = pd.read_csv("../data/pages-info.csv", dtype={"page_id": str})\
    [[ "page_id", "page_name", "fan_count" ]]
page_info.head()

sites = pd.read_csv(
    '../data/all-partisan-sites.csv',
    dtype={ "fb_id": str },
    na_values=["None"]
).rename(columns={"fb_id": "page_id"})

sites[[ "site", "political_category", "page_id",  ]].head()

Make sure that each page has been assigned only one political category:

assert (sites[
    ~sites["page_id"].isin([ "unavailable", "personal_page" ])
].groupby("page_id")["political_category"].nunique() > 1).sum() == 0

partisanship = sites.groupby("page_id")\
    ["political_category"].first()\
    .reset_index()

partisanship.head()

summary = pd.merge(
    page_info,
    partisanship,
    how="left"
)

summary.head()

summary["political_category"].value_counts()

summary.sort_values("page_id")\
    [["page_id", "political_category", "page_name", "fan_count" ]]\
    .to_csv("../output/fb-page-info-summary.csv", index=False)

---

---

---
