==> notebooks/01-aggregate-metrics.ipynb <==
# Calculate aggregate engagement metrics and get top posts

This notebook loads the raw data produced by Facebook's API, and then calculates the following: 

- Each page's number of posts and total engagement for each week and month between `2015-01-01` and `2017-03-31`, and for several specific time periods


- The total engagement for the 100 most popular posts on the left and right each week


- The top 50 posts overall, and by post type (photo, video, etc.)


import pandas as pd
from datetime import datetime, timedelta
from glob import glob

## Load the raw data

status_paths = glob("../data/statuses/*.csv")
len(status_paths)

def parse_status_csv(path):
    return pd.read_csv(
        path,
        parse_dates=['status_published'],
        date_parser=lambda x: pd.datetime.strptime(x, '%Y-%m-%d %H:%M:%S'),
    ).assign(page_id=lambda x: x["status_id"].str.split("_").str[0])

_statuses = pd.concat([ parse_status_csv(path) for path in status_paths ])\
    .drop_duplicates(subset=["status_id"])

len(_statuses)

## Select only statuses posted between `2015-01-01` and `2017-03-31`, and remove cover photos

(Changes to a page's cover photo are counted as posts in Facebook's API data.)

cover_photos = _statuses[
    (_statuses["status_type"] == "photo") &
    _statuses["status_link"].str.contains(r"^https://www.facebook.com") &
    _statuses["link_name"].str.contains(r" cover photo$")
]
len(cover_photos)

cover_photos["link_name"].value_counts().head()

to_date = lambda x: datetime.strptime(x, "%Y-%m-%d").date()

statuses = _statuses[
    ~_statuses["status_id"].isin(set(cover_photos["status_id"])) &
    (_statuses["status_published"].dt.date >= to_date("2015-01-01")) &
    (_statuses["status_published"].dt.date < to_date("2017-04-01"))
]

len(statuses)

## Add partisanship categorizations

page_info = pd.read_csv(
    '../output/fb-page-info-summary.csv',
    dtype={ "page_id": str },
)

statuses = pd.merge(
    page_info[[ "page_id", "page_name", "political_category" ]],
    statuses,
    on="page_id",
    how="left"
)

# Should be same number as before.
len(statuses)

### Pages with missing `political_category`

(There should be none.)

print("\n".join(statuses[
    statuses["political_category"].isnull()
]["page_id"].unique()))

## Calculate  `total_engagement` for each post

METRIC_COLUMNS = [ c for c in statuses.columns if "num_" in c ]

statuses[METRIC_COLUMNS] = statuses[METRIC_COLUMNS].astype(int)

statuses["total_engagement"] = statuses[[ "num_reactions", "num_comments", "num_shares" ]].sum(axis=1)

## Calculate metrics by week and month

def aggregate(statuses, groups, metric_fn=sum):
    grp = statuses.groupby(groups)
    df = grp[["total_engagement"] + METRIC_COLUMNS ].agg(metric_fn)
    df["post_count"] = grp.size()
    return df

def aggregate_pages_by_time(statuses, frequency, metric_fn=sum):
    time_grouper = pd.Grouper(
        key="status_published",
        freq=frequency,
        closed="left",
        label="left"
    )
    return aggregate(statuses, [ "page_id", time_grouper ], metric_fn)

### Weekly

weekly_totals = aggregate_pages_by_time(statuses, "W-SUN").astype(int)
weekly_totals.head()

weekly_totals.to_csv("../output/page-engagement-by-week.csv")

#### Just for Fox News and Occupy Democrats

fox_and_occupy_ids = page_info[
    page_info["page_name"].isin([ "Fox News", "Occupy Democrats" ])
].set_index("page_id")["page_name"].to_dict()
fox_and_occupy_ids

fox_vs_occupy_weekly_engagement = weekly_totals.loc[list(fox_and_occupy_ids.keys())]\
    ["total_engagement"].unstack(level=0)\
    .rename(columns=fox_and_occupy_ids).astype(int)

fox_vs_occupy_weekly_engagement.head()

fox_vs_occupy_weekly_engagement.to_csv("../output/fox-vs-occupy-engagement-by-week.csv")

### Monthly

monthly_totals = aggregate_pages_by_time(statuses, "MS").astype(int)
monthly_totals.head()

monthly_totals.to_csv("../output/page-engagement-by-month.csv")

## Get top 100 posts by week and partisanship

def get_top_100_posts_total_engagement(df):
    top = df.nlargest(100, "total_engagement")
    page_counts = top["page_id"].value_counts()
    total_engagement = top["total_engagement"].sum()
    return pd.Series({
        "num_posts": len(top),
        "total_engagement": total_engagement,
        "max_engagement": top["total_engagement"].max(),
        "min_engagement": top["total_engagement"].min(),
        "distinct_pages": len(page_counts),
        "top_page_count": page_counts.iloc[0],
        "top_page_prop_engagement": top.groupby("page_id")["total_engagement"].sum().max()\
            / total_engagement,
    })

weekly_top_posts_by_partisanship = statuses.groupby([
    pd.Grouper(
        key="status_published",
        freq="W-SUN",
        closed="left",
        label="left"
    ),
    "political_category",
]).apply(get_top_100_posts_total_engagement).unstack()

weekly_top_posts_by_partisanship.head()

weekly_top_posts_by_partisanship[[
    "distinct_pages",
    "top_page_count",
    "top_page_prop_engagement",
    "min_engagement"
]].describe()

# Make sure that we have at least 100 posts per week for both left/right
assert (weekly_top_posts_by_partisanship["num_posts"] < 100).sum().sum() == 0

weekly_top_posts_by_partisanship["total_engagement"]\
    .astype(int)\
    .to_csv("../output/top-100-posts-total-engagement-by-week-and-partisanship.csv")

## Get top 50 posts, overall, by post type


def get_top(df, num):
    return df.sort_values([
        "total_engagement",
        "status_id"
    ], ascending=False).head(num)

for status_type, group in statuses.groupby([ "status_type" ]):
    print(status_type)
    get_top(group, 50).to_csv(
        "../output/top-50-posts-{}.csv".format(status_type),
        index=False
    )

### ... and among all posts

get_top(statuses, 50).to_csv("../output/top-50-posts-all.csv", index=False)

---

---

---
