==> notebooks/02-calculate-page-growth.ipynb <==
# Calculate page growth, per engagement

This notebook calculates each pages' growth, in terms of changes in monthly engagement between March 2016 and March 2017.

import pandas as pd
import datetime

## Load the aggregated data

monthly_engagement = pd.read_csv("../output/page-engagement-by-month.csv", dtype={"page_id": str})

## Merge the two comparison months

comparison = pd.merge(
    monthly_engagement[
        monthly_engagement["status_published"] == "2016-03-01"
    ][[ "page_id", "total_engagement" ]],
    monthly_engagement[
        monthly_engagement["status_published"] == "2017-03-01"
    ][[ "page_id", "total_engagement" ]],
    on="page_id",
    how="outer",
    suffixes=[ ".march2016", ".march2017" ]
)
comparison.head()

## Add political category and page name information

page_info = pd.read_csv(
    "../output/fb-page-info-summary.csv",
    dtype={"page_id": str}
)
page_info.head()

def categorize_growth(page):
    diff = page["total_engagement.march2017"] - page["total_engagement.march2016"]
    if diff > 0:
        return "grown_{political_category}".format(**page)
    elif diff == 0:
        return "flat_{political_category}".format(**page)
    else:
        return "shrunken_{political_category}".format(**page)

comparison_with_fb_info = pd.merge(
    comparison,
    page_info,
    on="page_id",
    how="left"
).assign(growth_category=lambda x: x.fillna(0).apply(categorize_growth, axis=1))
comparison_with_fb_info.head()

# Make sure we have partisanship categorizations for each page
assert comparison_with_fb_info["political_category"].isnull().sum() == 0

comparison_with_fb_info.to_csv("../output/page-engagement-growth.csv", index=False)

## Pare down growth to top 20 pages on left and right  for graphics
This part of the notebook takes the top 20 pages on the left and right and pares down the growth spreadsheet to those top 20.

top_20s = comparison_with_fb_info.groupby("political_category")\
    .apply(lambda x: x.nlargest(20, "fan_count"))
top_20s

top_20s.loc["left"].to_csv("../output/page-engagement-growth-top-20-pages-left.csv", index=False)
top_20s.loc["right"].to_csv("../output/page-engagement-growth-top-20-pages-right.csv", index=False)

---

---

---
