==> notebooks/03-analyze-whois-registration-dates.ipynb <==
# Analysis of sites' WHOIS registration dates

The notebook below collects and analyzes the registration dates of the website domains for this project's list of sites.

The first section standardizes the site URLs to generate a list of unique domains, excluding domains belonging to blogging platforms.

The second section aggregates the WHOIS data for those domains by year, month, and political affiliation.

import pandas as pd
import re

## Creating the list of domains

### Loading the sites and converting URLs to domains

def site_to_domain(site):
    if pd.isnull(site): return None
    if re.match(r"^http", site):
        site = re.search("https?://([^/]+)", site).group(1)
    site = re.sub(r"/.*$", "", site)
    return ".".join(site.split(".")[-2:]).lower()

all_sites = pd.read_csv("../data/all-partisan-sites.csv")
all_sites["site"] = all_sites["site"].str.strip().str.lower()
all_sites["domain"] = all_sites["site"].apply(site_to_domain)
all_domains = all_sites["domain"].dropna().unique()
len(all_domains)

### Ignoring three domains from the list that belong to blogging platforms

DOMAINS_TO_IGNORE = [
    "blogspot.com",
    "blogspot.ca",
    "wordpress.com",
]

all_sites[
    all_sites["domain"].isin(DOMAINS_TO_IGNORE)
][[ "site", "domain", "political_category" ]]

### Generating a list of unique domains

unique_domains = list(sorted(set(all_domains) - set(DOMAINS_TO_IGNORE)))
unique_domains[:5]

with open("../output/unique-domains.txt", "w") as f:
    f.write("\n".join(unique_domains))

## Fetching WHOIS data for the domains

The domains above were then submitted to DomainTools' [Bulk Parsed Whois service](https://research.domaintools.com/bulk-parsed-whois/), which produced the file at [`data/domaintools-whois-results.csv`](../data/domaintools-whois-results.csv).

## Analyzing the domain registration dates

The main columns of interest in the WHOIS data are the domain and the WHOIS registration date:

whois = pd.read_csv("../data/domaintools-whois-results.csv", parse_dates=[ "create date" ])\
    .rename(columns={
        "create date": "date_registered"
    })
whois.head()[[ "domain", "date_registered" ]]

### Merging the WHOIS data with the manually-collected political affiliations

registration_dates = pd.merge(
    all_sites[[ "site", "domain", "political_category" ]],
    whois[[ "domain", "date_registered" ]],
    on="domain",
    how="left"
).drop_duplicates(subset=["site"])\
    .sort_values([ "date_registered", "site" ])

registration_dates["year_registered"] = registration_dates["date_registered"].dt.strftime("%Y")
registration_dates["month_registered"] = registration_dates["date_registered"].dt.strftime("%Y-%m")
registration_dates["date_registered"] = registration_dates["date_registered"].dt.strftime("%Y-%m-%d")

registration_dates = registration_dates.replace("NaT", pd.np.nan)

registration_dates.head()

registration_dates.to_csv("../output/whois-registration-dates.csv", index=False)

### Checking for sites for which we don't have WHOIS registration dates

Good sign: They are only those for which we have, above, explictly excluded the domains.

registration_dates[
    registration_dates["date_registered"].isnull()
]

### Aggregating registration dates by year, month, and political affiliation

annual_registrations = registration_dates\
    .dropna(subset=["date_registered"])\
    .groupby([ "year_registered", "political_category" ])\
    .size()\
    .unstack().fillna(0).astype(int)\
    .assign(total=lambda x: x.sum(axis=1))

annual_registrations.to_csv("../output/whois-registration-counts-annual.csv")
annual_registrations.head()

monthly_registrations = registration_dates\
    .dropna(subset=["date_registered"])\
    .groupby([ "month_registered", "political_category" ])\
    .size()\
    .unstack().fillna(0).astype(int)\
    .assign(total=lambda x: x.sum(axis=1))
    
monthly_registrations.to_csv("../output/whois-registration-counts-monthly.csv")
monthly_registrations.head()

---

---

---
